Fixes error when switching from random to black-10
When the random colors are present on the grid and you switched to
the 10% black setting, the random colors would not react correctly when
they were colored with the 10% black option. This update makes it so
that the random colors go straight to the 10% black color when they are
first painted over.